I edited my divs inside of flex-container
index.html:
I wanted my logo to be on the left side so I removed it from 'flex-container' as in css the style
was justify-content: flex-end, which made it go to the end.

styles.css:
I changed the 'margin' attribute to 'gap' attribute as I wanted to experiment with the 'gap' attribute.