Flexbox navbar finished
Html:
I made a main flex-container 'container', which fits an image and anothe
r flex-container 'login,about,menu'.

Css:
Style changes:
I changed the margin so that the image and font wouldn't be right at the
end of the viewport.
I changed the gap so that 'login,about,menu' won't be right next to each
other, which makes it hard to read.
I changed the background colour to match TOP's.
I changed the 'login,about,menu' font to white to make it easier to read

Note that you cannot use justify-content on the child container 'login,
about,menu' because the flex-container is only as large as the length of
words, which means that it won't go to the end of the viewport. Try and
adjust the position of the child elements directly from the grandparent/
parent (container), this will adjust all child elements.