css flexbox editing
I gave the main div a flex-attribute and gave it the justify-content of
'space-around', which means that it won't have space before and after th
e main content.

I also gave the class 'paragraph' a display of 'flex' so that I can work
with them seperate to the image(a different class).